% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_modals_verbs.R
\name{block_adj_prep_adv_fr}
\alias{block_adj_prep_adv_fr}
\title{Extract adjective, preposition, and adverb features}
\usage{
block_adj_prep_adv_fr(
  tokens,
  doc_ids,
  dict_lookup,
  word_lists_lookup,
  negation_adverbs
)
}
\arguments{
\item{tokens}{Annotated token data frame}

\item{doc_ids}{Document IDs}

\item{dict_lookup}{Dictionary lookup}

\item{word_lists_lookup}{Word lists lookup}

\item{negation_adverbs}{Vector of negation adverbs}
}
\value{
Data frame with f_39_prepositions through f_42_adverbs
}
\description{
Extract adjective, preposition, and adverb features
}
\keyword{internal}
