% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_tense_pronouns.R
\name{block_aux_tense_fr}
\alias{block_aux_tense_fr}
\title{Extract tense and auxiliary-related features}
\usage{
block_aux_tense_fr(tokens, doc_ids, head_lookup, proverb_pronouns)
}
\arguments{
\item{tokens}{Annotated token data frame}

\item{doc_ids}{Document IDs}

\item{head_lookup}{Head token lookup table}

\item{proverb_pronouns}{Vector of proverb pronoun lemmas}
}
\value{
Data frame with f_02_perfect_aspect and f_12_proverb_do
}
\description{
Extract tense and auxiliary-related features
}
\keyword{internal}
