% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_coordination_negation.R
\name{block_contractions_fr}
\alias{block_contractions_fr}
\title{Extract contraction features}
\usage{
block_contractions_fr(tokens, doc_ids)
}
\arguments{
\item{tokens}{Annotated token data frame}

\item{doc_ids}{Document IDs}
}
\value{
Data frame with f_59_contractions
}
\description{
Extract contraction features
}
\keyword{internal}
