% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_coordination_negation.R
\name{block_lexical_membership_fr}
\alias{block_lexical_membership_fr}
\title{Extract lexical membership features}
\usage{
block_lexical_membership_fr(tokens, doc_ids, word_lists_lookup)
}
\arguments{
\item{tokens}{Annotated token data frame}

\item{doc_ids}{Document IDs}

\item{word_lists_lookup}{Word lists lookup}
}
\value{
Data frame with f_10_demonstrative_pronoun through f_51_demonstratives
}
\description{
Extract lexical membership features
}
\keyword{internal}
