% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_modals_verbs.R
\name{block_modals_fr}
\alias{block_modals_fr}
\title{Extract modal verb features}
\usage{
block_modals_fr(tokens, doc_ids, dict_lookup)
}
\arguments{
\item{tokens}{Annotated token data frame with context columns}

\item{doc_ids}{Document IDs}

\item{dict_lookup}{Dictionary lookup}
}
\value{
Data frame with f_52_modal_possibility through f_54_modal_predictive
}
\description{
Extract modal verb features
}
\keyword{internal}
