% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_coordination_negation.R
\name{block_negation_fr}
\alias{block_negation_fr}
\title{Extract negation features}
\usage{
block_negation_fr(
  tokens,
  doc_ids,
  neg_synthetic_terms,
  negation_part_lemmas,
  negation_adverbs
)
}
\arguments{
\item{tokens}{Annotated token data frame}

\item{doc_ids}{Document IDs}

\item{neg_synthetic_terms}{Synthetic negation term lemmas}

\item{negation_part_lemmas}{Negation particle lemmas}

\item{negation_adverbs}{Negation adverb lemmas}
}
\value{
Data frame with f_66_neg_synthetic and f_67_neg_analytic
}
\description{
Extract negation features
}
\keyword{internal}
