% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_subordination.R
\name{block_participial_clauses_fr}
\alias{block_participial_clauses_fr}
\title{Extract participial clause features}
\usage{
block_participial_clauses_fr(tokens, doc_ids)
}
\arguments{
\item{tokens}{Annotated token data frame with participle markers}

\item{doc_ids}{Document IDs}
}
\value{
Data frame with f_24_infinitives through f_28_present_participle_whiz
}
\description{
Extract participial clause features
}
\keyword{internal}
