% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_passive.R
\name{block_passive_voice_fr}
\alias{block_passive_voice_fr}
\title{Extract passive voice and stative verb features}
\usage{
block_passive_voice_fr(tokens, doc_ids, head_lookup, passive_rel_values)
}
\arguments{
\item{tokens}{Annotated token data frame}

\item{doc_ids}{Document IDs}

\item{head_lookup}{Head token lookup table}

\item{passive_rel_values}{Dependency relation values indicating passive voice}
}
\value{
Data frame with f_17_agentless_passives, f_18_by_passives, f_19_be_main_verb, f_20_existential_there
}
\description{
Extract passive voice and stative verb features
}
\keyword{internal}
