% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_subordination.R
\name{block_relatives_fr}
\alias{block_relatives_fr}
\title{Extract relative clause features}
\usage{
block_relatives_fr(tokens, doc_ids)
}
\arguments{
\item{tokens}{Annotated token data frame with relative pronouns marked}

\item{doc_ids}{Document IDs}
}
\value{
Data frame with f_29_that_subj through f_34_sentence_relatives
}
\description{
Extract relative clause features
}
\keyword{internal}
