% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_modals_verbs.R
\name{block_specialized_verbs_fr}
\alias{block_specialized_verbs_fr}
\title{Extract specialized verb class features}
\usage{
block_specialized_verbs_fr(tokens, doc_ids, dict_lookup)
}
\arguments{
\item{tokens}{Annotated token data frame}

\item{doc_ids}{Document IDs}

\item{dict_lookup}{Dictionary lookup}
}
\value{
Data frame with f_55_verb_public through f_58_verb_seem
}
\description{
Extract specialized verb class features
}
\keyword{internal}
