% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_coordination_negation.R
\name{block_split_coordination_fr}
\alias{block_split_coordination_fr}
\title{Extract coordination features}
\usage{
block_split_coordination_fr(
  tokens,
  doc_ids,
  token_lookup,
  subject_heads,
  head_lookup,
  negation_part_lemmas
)
}
\arguments{
\item{tokens}{Annotated token data frame}

\item{doc_ids}{Document IDs}

\item{token_lookup}{Token lookup table}

\item{subject_heads}{Subject head lookup}

\item{head_lookup}{Head token lookup}

\item{negation_part_lemmas}{Negation particle lemmas}
}
\value{
Data frame with f_63_split_auxiliary, f_64_phrasal_coordination, f_65_clausal_coordination
}
\description{
Extract coordination features
}
\keyword{internal}
