% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_extraction.R
\name{dictionary_to_lemmas}
\alias{dictionary_to_lemmas}
\title{Extract lemmas from a dictionary entry}
\usage{
dictionary_to_lemmas(dict_lookup, feature, head_word = FALSE)
}
\arguments{
\item{dict_lookup}{The dict object}

\item{feature}{Feature name}
}
\value{
Character vector of lemmas
}
\description{
Extract lemmas from a dictionary entry
}
\keyword{internal}
