% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_extraction.R
\name{get_word_list}
\alias{get_word_list}
\title{Get a named word list from the word_lists data}
\usage{
get_word_list(word_lists_lookup, name)
}
\arguments{
\item{word_lists_lookup}{The word_lists object}

\item{name}{Name of the list to retrieve}
}
\value{
Character vector of terms
}
\description{
Get a named word list from the word_lists data
}
\keyword{internal}
