% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobibeR.fr-package.R
\docType{package}
\name{pseudobibeR.fr}
\alias{pseudobibeR.es}
\alias{pseudobibeR.es-package}
\alias{pseudobibeR.fr}
\title{pseudobibeR.fr: French Biber Feature Extraction}
\description{
Tools for computing multidimensional register features in French,
complementing the English-focused \code{pseudobibeR} package. The current
development iteration focuses on building language resources (dictionaries,
word lists) and parsing helpers grounded in Universal Dependencies.
}
\author{
\strong{Maintainer}: Maximiliano Cordovez \email{maximiliano.cf@outlook.es}

}
\keyword{internal}
