% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_extraction.R
\name{replace_nas}
\alias{replace_nas}
\title{Replace NAs with zeros in numeric columns of a data frame}
\usage{
replace_nas(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
Data frame with NAs in numeric columns replaced by 0
}
\description{
Replace NAs with zeros in numeric columns of a data frame
}
\keyword{internal}
